moved the rpc call dispatching to rpc_handler.rs and the function implementation to rpc_functions.rs. Added semi structured event call logging and track call durations.
