+++ src/lib.rs
@@ -3,3 +3,5 @@ pub mod io;
 pub mod safe_db;
 pub mod server;
 pub mod types;
+pub mod rpc_handler;
+pub mod rpc_functions;
